\name{MKmisc-package}
\alias{MKmisc-package}
\alias{MKmisc}
\docType{package}
\title{
Miscellaneous Functions from M. Kohl.
}
\description{
Miscellaneous Functions from M. Kohl.
}
\details{
\tabular{ll}{
Package: \tab MKmisc\cr
Type: \tab Package\cr
Version: \tab 0.7\cr
Date: \tab 2010-05-20\cr
Depends: \tab R(>= 2.10.0), stats, graphics, robustbase, RColorBrewer\cr
Suggests: \tab gplots\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
require(MKmisc)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
