\name{MKmisc-package}
\alias{MKmisc-package}
\alias{MKmisc}
\docType{package}
\title{
Miscellaneous Functions from M. Kohl.
}
\description{
Contains several functions for statistical data analysis; e.g. for sample size
and power calculations, computation of confidence intervals, and generation of
similarity matrices.
}
\details{
\tabular{ll}{
Package: \tab MKmisc\cr
Type: \tab Package\cr
Version: \tab 1.9\cr
Date: \tab 2022-11-19\cr
Depends: \tab R(>= 3.5.0)\cr
Imports: \tab stats, utils, graphics, grDevices, RColorBrewer, robustbase, ggplot2, scales, limma\cr
Suggests: \tab gplots, Amelia, knitr, rmarkdown, exactRankTests, foreach, parallel, doParallel\cr
License: \tab LGPL-3\cr
URL: \tab https://github.com/stamats/MKmisc\cr
}
library(MKmisc)
}
\author{ Matthias Kohl \url{https://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
