\name{MKmisc-package}
\alias{MKmisc-package}
\alias{MKmisc}
\docType{package}
\title{
Miscellaneous Functions from M. Kohl.
}
\description{
Contains several functions for statistical data analysis; e.g. for sample size
and power calculations, computation of confidence intervals, and generation of
similarity matrices.
}
\details{
\tabular{ll}{
Package: \tab MKmisc\cr
Type: \tab Package\cr
Version: \tab 0.996\cr
Date: \tab 2016-12-07\cr
Depends: \tab R(>= 2.14.0)\cr
Imports: \tab stats, graphics, grDevices, RColorBrewer, robustbase\cr
Suggests: \tab gplots, Amelia\cr
License: \tab LGPL-3\cr
URL: \tab http://www.stamats.de/\cr
}
library(MKmisc)
}
\author{ Matthias Kohl \url{http://www.stamats.de}

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
