\name{mkmisc-package}
\alias{mkmisc-package}
\alias{mkmisc}
\docType{package}
\title{
A package containing miscellaneous data analyses functions developed by Marcin Kierczak.
}
\description{
This package is a collection of various data analysis and visualization functions I have developed in my work. Primarily these are functions used in computational genetics.
}
\details{
\tabular{ll}{
Package: \tab mkmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-20\cr
License: \tab GPL (>= 2)\cr
}
Currently the package contains the following functions:\cr
plotMnhattanLD - plot linkage disequilibrium on your Manhattan plot
}
\author{
Marcin Kierczak

Maintainer: Marcin Kierczak <marcin.kierczak@slu.se>
}
\references{
}

\keyword{ package }
\seealso{
}

\examples{
}
